Added new `Status` section of the ResourcePool CRD
AFAICT the ResourceDemandStatus is not used anywhere anyway.
This adds a new `Status` section for the same sort of human-readabled
data.

I think in the future we'll have better and more machine-readable fields
to make it easier for capacity controllers to "see" how their request is
doing, but I think this is an ok start for now.